⚡ :: Integer를 Long으로 변경

+++ src/main/java/com/walkhub/walkhub/domain/notification/domain/NotificationListId.java
@@ -11,8 +11,8 @@ import java.io.Serializable;
 @EqualsAndHashCode
 public class NotificationListId implements Serializable {
 
-    private Integer notificationId;
+    private Long notificationId;
 
-    private Integer userId;
+    private Long userId;
 
 }
